+++ NEWS.md
@@ -1,5 +1,13 @@
 # Version 0.2.99.9000
 
+* A memoised function now has the same interface as the original function,
+  if the original function is known when `memoise` is called. (Otherwise,
+  the old behavior is invoked, with a warning.) (#14, @krlmlr)
+* The enclosing environment of the memoised function is specified explicitly,
+  defaults to `parent.frame()`.
+* `is.memoised` now checks if the argument is a function.
+* Testing infrastructure, full test coverage.
+
 # Version 0.2.1
 
 * Update to fix outstanding R CMD check issues.